-use-existing-mutate-out-path param was implemented
This setting allows to specify a path, from which PDBs are loaded by the mutagenesis module instead of being created by mut module.
Added also CPU setting 'MAX' and excluded assignment and parsing of ligand dock protocol to point in validation process, where it is done anyway. So it is not parsed over and over again.