Cleaned out-commented code from ScoreByPyRosetta, deleted FetchAndClean.py, Module-Params Documentation of ScoreByPyRosetta
Removed old, out-commented code about the use of InterFaceScoreCalculator since it is no longer in use.
Added also documentation print outs of ScoreByPyRosetta and some missing docstring information.
Renamed "calculate_fitness" to "calculate_fitness_score", since this makes it clearer, that this is a fitness determining score.
FetchAndClean.py is and was never really in use, since this process does not take much time if doing maually.